Moving lots of stuff into new blob/ folder.
Naming is to suggest that these things aren't necessarily
useful. Many are, but I'm not going to be supporting them as
I move forward.

Functionality either is or will be duplicated/improved upon in
PupTent, which is a bit more opinionated about how it is used
and therefore more generous in the returns it give to users.
The entity system that it's built around is allowing me to think
about each piece in a way that is closer to my ideal. This is
especially apparent in the case of the Locus and the relationship
of geometry expansion tools to the renderer(s).