Only use emplace_back when forwarding a new object
RenderSystem was often crashing after a call to
emplace_back that used an object that continues
to be referenced by its original source.
Not seeing the crash since switching to push_back.
We still use emplace_back for objects that are
created in-place.
Worth considering moving forward, so as not to adopt
move-like behavior willy-nilly.

Some trouble with script permissions.